Removed old event hooks when cleaning up the Slic engine (weird Slic errors with the Alexander scenario). Prevented compiler warnings and asserts.
Cleaned up some mod stuff that accidentally was placed in the main data files.
Updated credits to include zlib.

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@610 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gs/slic/SlicNamedSymbol.cpp
@@ -41,7 +41,7 @@
 namespace
 {
 //  Some value that is different from k_NORMAL_FILE (0) and k_TUTORIAL_FILE (1).
-    uint8 const k_GENERATED_BY_EXECUTABLE   = -1;
+    uint8 const k_GENERATED_BY_EXECUTABLE   = static_cast<uint8>(-1);
 }
 
 SlicNamedSymbol::SlicNamedSymbol(const char *name, SLIC_SYM type) :
